moves new player out of video-test
