+++ src/components/player/cue-bar.tsx
@@ -82,7 +82,17 @@ const MutePopupButton: React.FC<any> = () => {
       className="text-gray-400 rounded flex-nowrap flex items-center text-xs"
       onClick={() => setPlayerPrefs({muteNotes: !muteNotes})}
     >
-      {muteNotes ? <IconVolumeOn /> : <IconVolumeOff />}
+      {muteNotes ? (
+        <>
+          <span className="pr-1">unmute notes</span>
+          <IconVolumeOff />
+        </>
+      ) : (
+        <>
+          <span className="pr-1">mute notes</span>
+          <IconVolumeOn />
+        </>
+      )}
     </button>
   )
 }
@@ -98,7 +108,7 @@ const NoteCue: React.FC<any> = ({
   const [visible, setVisible] = React.useState(false)
   const {muteNotes} = getPlayerPrefs()
 
- // console.log('muteNotes: ', muteNotes)
+  // console.log('muteNotes: ', muteNotes)
 
   useCue(cue, actions)
 
@@ -197,13 +207,13 @@ const IconVolumeOff: React.FC<any> = ({className}) => {
         strokeWidth="2"
         d="M5.586 15H4a1 1 0 01-1-1v-4a1 1 0 011-1h1.586l4.707-4.707C10.923 3.663 12 4.109 12 5v14c0 .891-1.077 1.337-1.707.707L5.586 15z"
         clipRule="evenodd"
-      ></path>
+      />
       <path
         strokeLinecap="round"
         strokeLinejoin="round"
         strokeWidth="2"
         d="M17 14l2-2m0 0l2-2m-2 2l-2-2m2 2l2 2"
-      ></path>
+      />
     </svg>
   )
 }
@@ -222,7 +232,7 @@ const IconVolumeOn: React.FC<any> = ({className}) => {
         strokeLinejoin="round"
         strokeWidth="2"
         d="M15.536 8.464a5 5 0 010 7.072m2.828-9.9a9 9 0 010 12.728M5.586 15H4a1 1 0 01-1-1v-4a1 1 0 011-1h1.586l4.707-4.707C10.923 3.663 12 4.109 12 5v14c0 .891-1.077 1.337-1.707.707L5.586 15z"
-      ></path>
+      />
     </svg>
   )
 }
